Stock Market Analysis
Stock Market Analysis for the years 2014, 2015, and 2016 written in VBA.  Program is split up into functions and subroutines all calling eachother to perform this Analysis.  A form was created to the end user would know that the program was processing and to please wait.  The program informs the user the analysis is complete and to hit ok.  To run this program the main subroutine to be called is Stock_Market_Workbook().  Screenshots of the various years are provided to show the results of the program.